backend/test/unit/modules/mapsmap/assets/add_problem_asset.spec.ts
import { AddProblemAsset } from '../../../../../src/app/modules/mapsmap/assets/add_problem_asset';

describe('AddProblemAsset', () => {
  let transactionAsset: AddProblemAsset;

	beforeEach(() => {
		transactionAsset = new AddProblemAsset();
	});

	describe('constructor', () => {
		it('should have valid id', () => {
			expect(transactionAsset.id).toEqual(1);
		});

		it('should have valid name', () => {
			expect(transactionAsset.name).toEqual('addProblem');
		});

		it('should have valid schema', () => {
			expect(transactionAsset.schema).toMatchSnapshot();
		});
	});

	describe('validate', () => {
		describe('schema validation', () => {
      it.todo('should throw errors for invalid schema');
      it.todo('should be ok for valid schema');
    });
	});

	describe('apply', () => {
    describe('valid cases', () => {
      it.todo('should update the state store');
    });

    describe('invalid cases', () => {
      it.todo('should throw error');
    });
	});
});
